--- assets/css/estilos_1.css
@@ -1,41 +1,0 @@
-@import url(http://fonts.googleapis.com/css?family=Roboto:500,300);
-
-* {margin: 0;padding: 0;font-family:'Roboto', sans-serif;border:0;-webkit-box-sizing: border-box;-moz-box-sizing: border-box;-ms-box-sizing: border-box;box-sizing: border-box;font-weight:300}
-
-
-img{border:0}
-body {text-align:center;background:#e9e9e9}
-
-.calendario_ajax{margin:10px auto;width:700px;position:relative}
-	.calendario_ajax h2{text-align:center;margin:0 0 10px 0;font-size:1.5em;color:#000}
-
-	.calendario_ajax table.calendario {margin:0;width:100%;border:solid #ccc;border-width:1px 0 0 1px}
-		.calendario th {border:0;font-weight:500;background:#666;color:white;padding:10px;font-size:0.85em;height:40px}
-		.calendario td{padding:10px;text-align:center;border:solid #ccc;border-width:0 1px 1px 0;white-space:pre-line;font-size:2em;width:14.28%;color:#666;position:relative;background:#fff;height:80px}
-	.calendario td.desactivada {background:#e9e9e9;}
-	.calendario td.hoy{font-weight:500}
-		a.agregar_evento{display:block;position:absolute;right:0;top:0; background:#e9e9e9 url(../img/add.png) 50% 50% no-repeat;border-radius:0 0 10px 10px;width:24px;height:24px;text-decoration:none}
-	.calendario td a{color:red;text-decoration:none;}
-
-	.verde,.ok{font-weight:500;background:green;color:#fff}
-	.rojo{font-weight:500;background:red;color:#fff}
-	
-	
-	.toggle {background:#fff;}
-		.siguiente,.anterior{color:red;text-decoration:none;line-height:40px;}
-
-	#mask {position:absolute;left:0;top:0;z-index:9000;background:#000;display:none;width:100%;height:100%;}
-
-	#respuesta_form,#respuesta{clear:both;padding:20px 0}
-  
-	.window {position:absolute;top:20px;right:20px;left:20px;bottom:20px;display:block;z-index:9999;padding:20px;background-color:#ffffff;overflow:auto}
-
-		.window h2{text-align:center;margin:0 0 10px 0;font-size:1.5em;color:#000}
-		.window a.close{position:absolute;right:0;top:0;padding:10px;background:#e9e9e9 url(../img/delete.png) 50% 50% no-repeat;text-decoration:none}
-		.window p{margin:0 0 5px 0;font-size:1em;border:1px solid #e9e9e9;text-align:left;padding:10px;clear:both;}
-		
-		.window form{padding:10px;clear:both;background:#e9e9e9;float:left;width:100%;position:relative}
-		.window input[type=text]{border:0;background:white;width:64%;padding:5px;float:left}
-		.window input[type=button]{border:0;width:35%;background:#ccc;float:right;padding:5px}
-		.window p img{vertical-align:middle;float:right;border:0;width:16px;height:16px}
-		label.error {position:absolute;left:13px;top:13px;padding:0;width:20px;height:20px;background:#fff url(../img/delete.png) 50% 50% no-repeat;text-decoration:none;overflow:hidden;text-indent:-999em}
