+++ assets/css/material-dashboard.css
@@ -26,6 +26,8 @@
 }
 
 
+
+
 .noUi-target,
 .noUi-target * {
   -webkit-touch-callout: none;
