+++ core/agenda_1.php
@@ -1,0 +1,334 @@
+<?php
+
+require_once '../classes/SearchFindShow.php';
+require_once '../classes/UserSessions.php';
+require_once '../classes/MyErrorHandler.php';
+
+session_start();
+
+/**
+ * 
+ * @Global Var
+ * 
+ */
+
+$correo = "";
+
+
+
+function SanityCheck(){
+    
+    global $correo;
+    
+if (isset($_SESSION['myemail'])){
+//desde sigin.php
+    
+    //echo "En sanity...";
+    
+    $correo = $_SESSION['myemail'];
+
+    return 1;
+
+    }
+        return 0;
+    }
+
+function redirect($url, $statusCode = 303)
+{
+   header('Location: ' . $url, true, $statusCode);
+   die();
+}
+
+
+//****
+//Valida la session contra Base de Datos
+
+function ShieldSession(){
+    
+    global $correo;
+        
+    //echo "Shield..";
+    
+      
+            $sess = new UserSessions(); 
+            //Echo "After Sanity...";
+            $Ecode =  $sess->CheckSessionInDb(session_id(),$correo);
+            //echo $Ecode;
+            
+        if ($Ecode!="302"){
+            
+            redirect("https://".$_SERVER['SERVER_NAME']."/corvi/core/acceso.php");
+            
+        
+        }
+      
+           
+      
+    
+    
+    
+}
+
+/**
+ * Valida si la variablle correo esta seteada como
+ * parte de la sesion global
+ * 
+ */
+if (SanityCheck()){
+    ShieldSession();
+    }
+else{
+    redirect("https://".$_SERVER['SERVER_NAME']."/corvi/core/acceso.php"); 
+}
+
+?>
+
+
+
+<!doctype html>
+<html lang="en">
+<head>
+	<meta charset="utf-8" />
+	<link rel="apple-touch-icon" sizes="76x76" href="../assets/img/apple-icon.png" />
+	<link rel="icon" type="image/png" href="../assets/img/favicon.png" />
+	<meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1" />
+
+	<title>Corvi</title>
+
+	<meta content='width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0' name='viewport' />
+    <meta name="viewport" content="width=device-width" />
+
+    <!-- Bootstrap core CSS     -->
+    <link href="../assets/css/bootstrap.min.css" rel="stylesheet" />
+
+    <!--  Material Dashboard CSS    -->
+    
+
+   <link type="text/css" rel="stylesheet" media="all" href="../assets/css/estilos.css">
+    
+   
+
+    <!--     Fonts and icons     -->
+    <link href="https://maxcdn.bootstrapcdn.com/font-awesome/latest/css/font-awesome.min.css" rel="stylesheet">
+    <link href='https://fonts.googleapis.com/css?family=Roboto:400,700,300|Material+Icons' rel='stylesheet' type='text/css'>
+    
+      
+
+       
+
+	<!-- helper libraries -->
+	
+        <script src="https://code.jquery.com/jquery-1.11.1.min.js"></script>
+        
+	<script src="https://ajax.aspnetcdn.com/ajax/jquery.validate/1.12.0/jquery.validate.min.js"></script>
+	<script src="https://ajax.aspnetcdn.com/ajax/jquery.validate/1.12.0/localization/messages_es.js "></script>
+        
+        
+        
+        <script>
+	function generar_calendario(mes,anio)
+	{
+		var agenda=$(".cal");
+		agenda.html("<img src='../assets/img/loading.gif'>");
+		$.ajax({
+			type: "GET",
+			url: "ajax_calendario.php",
+			cache: false,
+			data: { mes:mes,anio:anio,accion:"generar_calendario" }
+		}).done(function( respuesta ) 
+		{
+			agenda.html(respuesta);
+		});
+	}
+		
+	function formatDate (input) {
+		var datePart = input.match(/\d+/g),
+		year = datePart[0].substring(2),
+		month = datePart[1], day = datePart[2];
+		return day+'-'+month+'-'+year;
+	}
+		
+		$(document).ready(function()
+		{
+			/* GENERAMOS CALENDARIO CON FECHA DE HOY */
+			generar_calendario("<?php if (isset($_GET["mes"])) echo $_GET["mes"]; ?>","<?php if (isset($_GET["anio"])) echo $_GET["anio"]; ?>");
+			
+			
+			/* AGREGAR UN EVENTO */
+			$(document).on("click",'a.add',function(e) 
+			{
+				e.preventDefault();
+				var id = $(this).data('evento');
+				var fecha = $(this).attr('rel');
+				//var id_evento = $(this).data('evento');
+				$('#mask').fadeIn(1000).html(
+                                "<div id='nuevo_evento' class='window' rel='"+fecha+"'>Agendar una  visita para "+formatDate(fecha)+"</h2><a href='#' class='close' rel='"+fecha+"'>&nbsp;</a><div id='respuesta_form'></div><form class='formeventos'><input type='text' id='hora' name='hora' class='form-control' value='09:30:00'><input type='text' rutcom='rutcom' name='rutcom' class='form-control' value='xx.xxx.xxx-x'><input type='text' name='rolid' id='rolid' class='required' value='Rol de casa escogida'><input type='text' name='evento_titulo' id='evento_titulo' class='required'><input type='button' name='Enviar' value='Guardar' class='enviar'><input type='hidden' name='evento_fecha' id='evento_fecha' value='"+fecha+"'></form></div>"
+           
+                                );
+			});
+			
+			/* LISTAR EVENTOS DEL DIA */
+			$(document).on("click",'a.modal',function(e) 
+			{
+				e.preventDefault();
+				var fecha = $(this).attr('rel');
+				
+				$('#mask').fadeIn(1000).html("<div id='nuevo_evento' class='window' rel='"+fecha+"'>Eventos del "+formatDate(fecha)+"</h2><a href='#' class='close' rel='"+fecha+"'>&nbsp;</a><div id='respuesta'></div><div id='respuesta_form'></div></div>");
+				$.ajax({
+					type: "GET",
+					url: "ajax_calendario.php",
+					cache: false,
+					data: { fecha:fecha,accion:"listar_evento" }
+				}).done(function( respuesta ) 
+				{
+					$("#respuesta_form").html(respuesta);
+				});
+			
+			});
+		
+			$(document).on("click",'.close',function (e) 
+			{
+				e.preventDefault();
+				$('#mask').fadeOut();
+				setTimeout(function() 
+				{ 
+					var fecha=$(".window").attr("rel");
+					var fechacal=fecha.split("-");
+					generar_calendario(fechacal[1],fechacal[0]);
+				}, 500);
+			});
+		
+			//guardar evento
+			$(document).on("click",'.enviar',function (e) 
+			{
+				e.preventDefault();
+				if ($("#evento_titulo").valid()==true)
+				{
+					$("#respuesta_form").html("<img src='../assets/img/loading.gif'>");
+					var evento=$("#evento_titulo").val();
+					var fecha=$("#evento_fecha").val();
+                                        var hora=$("#hora").val();
+                                        var rut_com=$("#rut_com").val();
+                                        //var rut_ven=$("#rut_ven").val();
+                                        var rolid=$("#rolid").val();
+					$.ajax({
+						type: "GET",
+						url: "ajax_calendario.php",
+						cache: false,
+						data: { rut_com:rut_com,rolid:rolid,evento:evento,fecha:fecha,hora:hora,accion:"guardar_evento" }
+					}).done(function( respuesta2 ) 
+					{
+						$("#respuesta_form").html(respuesta2);
+						$(".formeventos,.close").hide();
+						setTimeout(function() 
+						{ 
+							$('#mask').fadeOut('fast');
+							var fechacal=fecha.split("-");
+							generar_calendario(fechacal[1],fechacal[0]);
+						}, 3000);
+					});
+				}
+			});
+				
+			//eliminar evento
+			$(document).on("click",'.eliminar_evento',function (e) 
+			{
+				e.preventDefault();
+				var current_p=$(this);
+				$("#respuesta").html("<img src='../assets/img/loading.gif'>");
+				var id=$(this).attr("rel");
+				$.ajax({
+					type: "GET",
+					url: "ajax_calendario.php",
+					cache: false,
+					data: { id:id,accion:"borrar_evento" }
+				}).done(function( respuesta2 ) 
+				{
+					$("#respuesta").html(respuesta2);
+					current_p.parent("p").fadeOut();
+				});
+			});
+				
+			$(document).on("click",".anterior,.siguiente",function(e)
+			{
+				e.preventDefault();
+				var datos=$(this).attr("rel");
+				var nueva_fecha=datos.split("-");
+				generar_calendario(nueva_fecha[1],nueva_fecha[0]);
+			});
+
+		});
+		</script>
+    
+    
+   
+    
+</head>
+
+<body>
+
+	<div class="wrapper">
+
+	    
+
+	    <div class="main-panel">
+                
+			
+                        
+                
+                
+                
+                <div class="content">
+                    <div id="header">
+			
+        </div>
+        <div class="shadow"></div>
+        <div class="hideSkipLink">
+        </div>
+        <div class="main">
+            <div class="calendario_ajax">
+		<div class="cal"></div><div id="mask"></div>
+	</div>
+            
+
+
+          
+            
+            
+
+        </div>
+        <div class="clear">
+        </div>
+      
+                </div>
+                
+
+                
+                
+                
+
+ 
+
+			
+
+
+
+
+
+			
+		</div>
+	</div>
+
+</body>
+
+	
+
+	
+
+	
+
+	
+        
+        
+
+</html>
